pyproject.toml
[tool.poetry]
name = "cytotools"
version = "0.1.0"
description = ""
authors = ["burtonrj <burtonrj@cardiff.ac.uk>"]

[tool.poetry.dependencies]
python = "^3.8"
tqdm = "^4.64.0"
ipython = "^8.4.0"
numpy = "^1.22.4"
pandas = "^1.4.2"
phate = "^1.0.8"
scikit-learn = "^1.1.1"
umap-learn = "^0.5.3"
detecta = "^0.0.5"
KDEpy = "^1.1.0"
scikit-fda = "^0.7.1"
alphashape = "^1.3.1"
Shapely = "^1.8.2"
polars = "^0.13.49"
FlowUtils = "^1.0.0"
FlowIO = "^1.0.1"
pyarrow = "^8.0.0"
s3fs = "^2022.5.0"
bson = "^0.5.10"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
Sphinx = "^5.0.2"
sphinx-rtd-theme = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
